BREAKING CHANGE: Use config for value determination
