--- p1_influx_db/dsmr_parse/__init__.py
@@ -1,2 +1,0 @@
-from .main import parse_dsmr_telegram
-from .formats import *
