--- p1_influx_db/dsmr_parse/formats.py
@@ -1,12 +1,0 @@
-from collections import namedtuple
-from pydantic import BaseModel, AwareDatetime
-from typing import Dict, Any
-
-dsmrMessages = namedtuple("dsmrMessages", ["topic", "payload"])
-
-
-class p1Messages(BaseModel):
-    measurement: str
-    tags: Dict[str, Any]
-    fields: Dict[str, Any]
-    time: AwareDatetime
