+++ p1_influx_db/example_config.toml
@@ -14,3 +14,32 @@ key = "etc/certificate/key"
 
 [p1]
 device = "/dev/serial0"
+
+[p1.electricity]
+keys = [ELECTRICITY_USED_TARIFF_1,
+        ELECTRICITY_USED_TARIFF_2,
+        ELECTRICITY_DELIVERED_TARIFF_1,
+        ELECTRICITY_DELIVERED_TARIFF_2]
+
+[p1.electricity_current]
+keys = [CURRENT_ELECTRICITY_USAGE,
+        CURRENT_ELECTRICITY_DELIVERY,
+        INSTANTANEOUS_ACTIVE_POWER_L1_POSITIVE,
+        INSTANTANEOUS_ACTIVE_POWER_L2_POSITIVE,
+        INSTANTANEOUS_ACTIVE_POWER_L3_POSITIVE,
+        INSTANTANEOUS_ACTIVE_POWER_L1_NEGATIVE,
+        INSTANTANEOUS_ACTIVE_POWER_L2_NEGATIVE,
+        INSTANTANEOUS_ACTIVE_POWER_L3_NEGATIVE]
+
+[p1.voltage]
+keys = [INSTANTANEOUS_VOLTAGE_L1,
+        INSTANTANEOUS_VOLTAGE_L2,
+        INSTANTANEOUS_VOLTAGE_L3]
+
+[p1.current]
+keys = [INSTANTANEOUS_CURRENT_L1,
+        INSTANTANEOUS_CURRENT_L2,
+        INSTANTANEOUS_CURRENT_L3]
+
+[p1.gas]
+keys = [HOURLY_GAS_METER_READING]
