+++ p1_influx_db/httpmain.py
@@ -1,27 +1,30 @@
+from html import parser
 from loguru import logger
-import toml
 import asyncio
+from typing import Dict, Any
 
+from dsmr_parse import dsmrParse
 from dsmr_parser import telegram_specifications
 from dsmr_parser.clients import AsyncSerialReader, SERIAL_SETTINGS_V5
-from .dsmr_parse import parse_dsmr_telegram
 from aiohttp.client_exceptions import ClientConnectorError
 
 
-async def parse_telegram_influx(name, queue: asyncio.Queue, config_file: str):
+async def parse_telegram_influx(
+    name, queue: asyncio.Queue, config_file: str, parser: dsmrParse
+):
     while True:
         telegram = await queue.get()
-        info_list = parse_dsmr_telegram(telegram)
+        info_list = parser.parse_dsmr_telegram(telegram)
         from .http_to_influxdb import send_parsed_telegram
 
         results = await send_parsed_telegram(info_list, queue, config_file)
         queue.task_done()
 
 
-async def httpmain(config_file="./p1_influx_db/config.toml"):
+async def mainhttp(config: Dict[str, Any]):
     queue = asyncio.Queue()
-    with open(config_file, "r") as f:
-        config = toml.load(f)
+
+    dsmrParser = dsmrParse(config["p1"])
     logger.info("Opening SerialReader")
     serial_reader = AsyncSerialReader(
         device=config["p1"]["device"],
@@ -30,7 +33,16 @@ async def httpmain(config_file="./p1_influx_db/config.toml"):
     )
     tasks = []
     for i in range(3):
-        tasks.append(asyncio.create_task(parse_telegram_influx(f"worker-{i}", queue, config_file=config_file)))
+        tasks.append(
+            asyncio.create_task(
+                parse_telegram_influx(
+                    f"worker-{i}",
+                    queue,
+                    config=config["p1"]["filepath"],
+                    parser=dsmrParser,
+                )
+            )
+        )
 
     await serial_reader.read_as_object(queue)
     result = await asyncio.gather(*tasks, return_exceptions=True)
