+++ p1_influx_db/mqttmain.py
@@ -1,20 +1,18 @@
 from loguru import logger
-import toml
+from typing import Dict, Any
 
 from dsmr_parser import telegram_specifications
 from dsmr_parser.clients import SerialReader, SERIAL_SETTINGS_V5
 
+from dsmr_parse import dsmrParse
 
 from p1_influx_db.mqtt import Client
-from p1_influx_db.dsmr_parse import parse_dsmr_telegram
 
 
-def mqttmain(config_file="./p1_influx_db/example_config.toml"):
+def mqttmain(config: Dict[str, Any]):
     logger.info("Opening SerialReader")
-    with open(config_file, "r") as f:
-        config = toml.load(f)
-        mqtt_config = config["mqtt"]
-
+    mqtt_config = config["mqtt"]
+    dsmrParser = dsmrParse(config["p1"])
     client = Client(**mqtt_config)
     client.start()
     serial_reader = SerialReader(
@@ -24,7 +22,7 @@ def mqttmain(config_file="./p1_influx_db/example_config.toml"):
     )
     try:
         for telegram in serial_reader.read_as_object():
-            info_list = parse_dsmr_telegram(telegram)
+            info_list = dsmrParser.parse_dsmr_telegram(telegram)
 
             for dsmr_message in info_list:
                 client.publish_messages(dsmr_message)
